client/bridge/bridge-python/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
PyJWT>=2.9.0,<3.0.0
